# Conflicts:
#	Block/Form/Cc.php
#	Controller/Return/Index.php
#	Helper/Config.php
#	Helper/Data.php
#	Helper/Vault.php
#	Model/Ui/AdyenCcConfigProvider.php
#	Model/Ui/AdyenOneclickConfigProvider.php
#	Observer/AdyenOneclickDataAssignObserver.php
#	etc/config.xml
#	etc/events.xml
#	etc/frontend/di.xml

+++ Api/Data/CreditmemoInterface.php
@@ -67,15 +67,15 @@ interface CreditmemoInterface
     public function setCreditmemoId(int $creditMemoId): CreditmemoInterface;
 
     /**
-     * @return int|null
+     * @return float|null
      */
-    public function getAmount(): ?int;
+    public function getAmount(): ?float;
 
     /**
-     * @param int $amount
+     * @param float $amount
      * @return CreditmemoInterface
      */
-    public function setAmount(int $amount): CreditmemoInterface;
+    public function setAmount(float $amount): CreditmemoInterface;
 
     /**
      * @return int|null
